// ClientApp/src/app/shared/models/customerAppointment.model.ts
export interface CustomerAppointment {
  statusDescription: string;
  employeeName: string;
  customerName: string;
  serviceDescription: string;
  appointmentShopDescription: string;
  areaCode: string;
  areaDescription: string;
  regionCode: string;
  regionDescription: string;
  appointmentID: string;
  statusCode: string;
  appointmentDate: string;
  duration: number;
  employeeCode: string;
  customerCode: string;
  appointmentShopCode: string;
  serviceCode: string;
  note: string;
  roomCode: string;
  rowGuid: string;
  shopDescription?: string;
	roomDescription?: string;
	mediaTypeDescription?: string;
}